tests/wpt/web-platform-tests/app-history/navigate/goTo-after-adding-iframe.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
async_test(t => {
  let start_length = history.length;

  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  window.onload = () => t.step_timeout(() => {
    appHistory.navigate("#top");
    assert_equals(appHistory.entries().length, 2);

    let i = document.createElement("iframe");
    i.src = "/common/blank.html";
    document.body.appendChild(i);
    i.onload = () => t.step_timeout(() => {
      assert_equals(i.contentWindow.appHistory.entries().length, 1);
      i.contentWindow.appHistory.navigate("#fragment");
      assert_equals(i.contentWindow.appHistory.entries().length, 2);
      i.contentWindow.appHistory.back().then(t.step_func(() => {
        assert_equals(i.contentWindow.appHistory.entries().length, 2);
        assert_equals(i.contentWindow.appHistory.current.index, 0);
        return i.contentWindow.appHistory.forward();
      })).then(t.step_func_done(() => {
        assert_equals(i.contentWindow.appHistory.entries().length, 2);
        assert_equals(i.contentWindow.appHistory.current.index, 1);
      }));
    }, 0);
  }, 0);
}, "appHistory.goTo() should work in an iframe that is not present in all history entries");
</script>
</body>
